+++ src/lib/index.ts
@@ -33,6 +33,7 @@ export {
 	removeFiles,
 	fail,
 	type SuperValidated,
+	type SuperValidateOptions,
 	type TaintedFields,
 	type ValidationErrors,
 	type Infer,