+++ CHANGELOG.md
@@ -20,7 +20,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changed
 
-- Bumped TypeBox version requirement to 0.34.28.
+- Bumped TypeBox peerDep requirement to 0.34.28.
+- Bumped Effect peerDep requirement to 3.13.7.
+- Bumped Valibot peerDep requirement to 1.0.0-rc.3.
 
 ## [2.23.1] - 2025-01-21